# like-service/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Utilisateurs",
    "version": "1.0.0"
  },
  "paths": {
    "/likes": {
      "post": {
        "summary": "Ajouter un like à un post par un utilisateur",
        "tags": [
          "Likes"
        ],
        "requestBody": {
          "description": "ID utilisateur et ID du post à liker",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "postId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "608c1f2f9a1d3b0015a4b123"
                  },
                  "postId": {
                    "type": "string",
                    "example": "608c1f2f9a1d3b0015a4b456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Like créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Like"
                }
              }
            }
          },
          "400": {
            "description": "Like déjà existant ou données manquantes"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "get": {
        "summary": "Liste tous les likes (optionnel)",
        "tags": [
          "Likes"
        ],
        "responses": {
          "200": {
            "description": "Liste des likes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Like"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/likes/{id}": {
      "delete": {
        "summary": "Supprimer un like par son ID",
        "tags": [
          "Likes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID du like à supprimer",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Like supprimé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Like supprimé"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Like non trouvé"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Like": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60af8844c25e5b0022a7f5b2"
          },
          "userId": {
            "type": "string",
            "example": "608c1f2f9a1d3b0015a4b123"
          },
          "postId": {
            "type": "string",
            "example": "608c1f2f9a1d3b0015a4b456"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-30T10:30:00.000Z"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Likes",
      "description": "Gestion des likes"
    }
  ]
}